Extract sendstring into its own compilation unit
